Spline IK: Increase maximum allowable chain length
The maximum allowable spline IK chain length was increased from 255 to
4096, and the maximum preview and render U resolutions for curves and
surfaces was increased from 1024 to 8192. A higher U resolution allows
curves to more smoothly guide long spline IK chains.

The maximum allowable preview and render V resolutions for surfaces
was increased to 8192. Although surfaces are not used for spline IK,
the V resolution is used alongside the U resolution, and its maximum
allowable value was therefore increased to be the same as U for
parity.

Some array sizes in armature_update.c were increased to 4096 to
accommodate the larger allowable spline IK chains.

A value in overlay_armature.c was increased to 4096 to
accommodate the larger allowable spline IK chains.