Eevee-next: Lookdev HDRI Switcher
This PR adds support to override the world with the HDRI of the
viewport.

Blurring of the background and camera based rotation isn't supported.
Note the workflow when using irradiance caches isn't clear. As
irradiance caches are created without the viewport known it would
still bake the scene world in.

Taken that into account the HDRI switcher would only change the
reflections. We should find out how we could solve that.