Don't calculate homogenous offset if viewspace offset is zero
Users were reporting offset issues when the retopology overlay was disabled.
The reason those issues were happening is because of `vs_offset = min(vs_offset, vs_z * -0.5);`.
That line is necessary for proper functioning of the retopology overlay, but causes issues at lower offset values (such as zero, which the retopology overlay is disabled).
This is kind of a band-aid solution, but it's a solution nonetheless.