Always use zero for retopology min offset when disabled
Using 0.0015f as minimum value on Apple makes sense when the retopology overlay is enabled.
When disabled however, this will cause the shader to think the overlay is enabled when it's not, affecting the color of faces.
Therefore the offset when disabled should always be zero.
I've removed the unnecessary define and shortened the name of the other one.