Revert "Transform: Allow navigation by default"
This reverts commit d53862351d042c4c36c7ebb49feb230725e43a0b.

After testing artists at the studio, it was observed that changing the modal maps of Transform is quite disruptive because the features of "Proportional Editing" and "Automatic Constraint" are heavily used.

Therefore, it is interesting to allow the user to decide when or not to adapt muscle memory to new changes.