Fix: Adding node group asset doesn't respect import method
The asset import method option was added after the additions that use
assets in the node editor, so the node editor features still always used
append and reuse, no matter the import method.

This commit improves the asset import function to use the asset's
import method, with "Append & Reuse" as a default.