Fix #109570: Error when using collection parameters in drivers
This was unintentionally fixed in the main branch for the 4.0 series
in the 61ed5b3d24. Initially the commit seemed to not be leading to
any functional changes, but turns out it did fix a real use-case.

This commit is the minimalistic change extracted from the original
commit, making it safe to be ported to the 3.6 LTS.