UI: Make 3D Viewport header transparent
Back in Blender 3.0, the header and toolbar position were swapped to
follow hierarchy of settings. However, due to the header having usually
less buttons than the toolbar, it looked weird so it was decided to
make it fully opaque.

Based on user feedback it is preferred to have transparency and gain that
extra bit of space, even if at times it could look weird due to the header
having less items.

This PR makes it so the header supports region overlap, and makes it fully
transparent.


Fixes #92844

Related PR !107094