Fix #109679: Normal strength is applied incorrectly
Previously the normal strength calculations were linearly extrapolating
the supplied normal map respect to different scaling in the xy resp.
z-axis. This new method takes this into account and applies it in the
correct way.

The calculations are identical to the ones presented in #109679 but
translated into glsl.

Please see #109679 for more information and background.