Refactor: Mesh Merge: deduplicate code, no interpolation for edges
With 74772c6920, the merge by distance code for meshes now supports
optional interpolation of custom data for vertices.

As this can be advantageous for performance and memory, it seems
convenient to support the same for edges (and in the future polygons
and corners).

So this commit reworks the code that finds the edge groups to match the
same used for vertices and thus deduplicating and simplifying the code.

Conveniently the `edge_ctx_map` array is no longer needed and can be
removed to save memory.

However it was necessary to create another array (which is usually
smaller) called `double_edges` (to match `double_verts`).

---
The profiling result depends on some factors such as:
- interpolate or not interpolate customdata;
- affect a large or small portion of the mesh;

Testing required many repetitions as the values are quite close.
Although small, the difference is only seen in cases of
Non-interpolation that affects large portions of the mesh:

Does not interpolate + affects large portions of the high-poly mesh:
| Before: | After: | Difference factor: |
| 1297.64 ms | 1240.01 ms | +4.4% |

Interpolates + affects large portions of the high poly mesh
| Before: | After: | Difference factor: |
| 248.59 ms | 249.76 ms | -0.5% |

Does not Interpolate + affects small portions of the high poly mesh
| Before: | After: | Difference factor: |
| 664.52 ms | 654.28 ms | +1.5% |

Interpolates + affects small portions of the high poly mesh
| Before: | After: | Difference factor: |
| 1081.25 ms | 1089.4 ms | -0.1% |