WIP: UI: Consistent Menu/Block/Popup Ordering
All Content is shown in natural top-down order regardless of where it
is initiated.

---

Currently, Blender menus are what is called "directional", where menu items are sorted frop top-bottom or bottom-top based on the direction the menu open. It might be a nice time, with Blender 4.0, to remove this feature and have content that is always in a consistent order.

We have a preference for this behavior but haven't exposed it to users. But I don't think that we should enable this preference. If we (re)introduce this preference, make it off by default now then on by default later, we will be stuck with this preference forever. But whether on or off the description of this preference will never make sense because of how inconsistent we are with changing the direction of various items. With "Directional" turned off then all dropdown and popup content will always have the same order.

But now consider that "Directional" preference when off. Turning it ON will then mean that SOME content in some circumstances will reverse. So the 3D Viewport, in Object Mode, if you move the header to the bottom, of the four items in the center, just one of them - Transform Pivot Point - will reverse direction. Not sure how to explain that as an option.

Another reason is that we can consider changing the order of menu items if we have one direction rather than two. The change to consistent order makes relatively small changes except for the Timeline menus. For those menus everyone is used to a particular order because that editor is generally at the bottom. If we change to consistent ordering then we could also consider changing the Timeline menus item order.

RCS Suggestion: https://blender.community/c/rightclickselect/14vo