Mesh: Rename "polys" to "faces"
Implements part of #101689.

The "poly" name was chosen to distinguish the `MLoop` + `MPoly` system
from the `MFace` system it replaced. Those two structures persisted
together for a long time, but nowadays `MPoly` is gone, and `MFace`
is only used in some legacy code like the particle system.

To avoid unnecessarily using a different term, increase consistency with
the UI and with BMesh, and generally make code a bit easier to read,
this commit replaces the `poly` term with `poly`. Most variables that
use the term are renamed too. `Mesh.fdata` and `Mesh.totface` now
have a `_legacy` suffix to reduce confusion. In a next step, `pdata`
can be renamed to `face_data` as well.