Walk Navigation: Controls to adjust jump height interactively
Pressing '.' increases, ',' decreases the jump height in modal walk navigation.
Before, you could only change the global height in the preferences, this change
allows you to do so interactively (the setting in preferences is still loaded as default
when starting walk navigation). This is useful when different parts of the scene
have varying elevation changes.