UI: experimental support for "alert" property in menus/popups
The red alert color in the UI is currently only supported for:

- uiItemFullR
- uiItemFullO_ptr_ex
- uiItemL_

This patch adds experimental support for this for menus. The intention
behind this is to indicate an alert further down the menu on the menu/
popup itself, so it is more discoverable. Put this up to see if we even
want this at all (not sure if this is very common in other UIs tbh).

For testing purposes, some existing menus have been marked with the
alert property (this is to be removed prior to possible landing of this
patch).

![image](/attachments/b878c461-d4ed-430a-9561-94d88b0137f7)
![image](/attachments/9a791539-09f7-40ed-b5fd-dcd8dd729ce8)
![image](/attachments/5296195d-7b52-4750-9ff7-ef4ef27fcc32)

Fixes #109778.