Cleanup: Move definition of some navigation operators to 'view3d_ops.cc'
Since we are unifying the implementation of the navigation operators
and using them in a non-operator utility, they no longer have a strong
dependency on the operator API and therefore the definition of these
operators can be moved to `view3d_ops.cc` itself.

This more closely matches the design seen in the transform code.

There are still many other navigation operators that can be adapted.

Another change is that lambda has been used instead of static functions
for the operators' invoke callbacks.