Fix #109630: Accent Grave / Tilde key on macOS not working
Through trial and error, it was discovered that the Instance Property
`charactersIgnoringModifiers ` and the function `UCKeyTranslate` with
keyAction `kUCKeyActionDown`, do not detect the characters ` and '.

So just use `kUCKeyActionUp` in `UCKeyTranslate`.

Comments have been added to explain each decision.

---
Apparently `charactersByApplyingModifiers` could be used to bypass `charactersIgnoringModifiers`, but the lack of documentation on which modifiers are allowed in this method makes the usage unclear.