I18n: make the string holding light cache messages longer
The `light_cache_info` string holds information about the outcome of
Eevee light cache baking. It is currently 64 bits, which is enough to
hold all possible messages in English. However, some messages need to
be longer than this in other languages, so the holding string also
needs to be longer. From what I can tell, 128 bytes is plenty enough.