Simplified static types selection for node tree interface items.
The C++ static dispatch methods (to_static_type like CPPType) are nice
for "true" C++ classes, but when mixing with C structs (DNA) it becomes
a headache to manage fake inheritance and similar features.

Better to keep it simple and stupid for now, even if that means
having centralized functions that have to handle all the types.