WIP: Initial implementation of new blendfile compatibility policy.
This PR makes the following changes:
* The minfileversion stored in Blender files is now a hard limit,
  Blender does not open any file which minfileversion is newer than its
  current version.
** An error message is shown to user when trying to open such file.
** When trying to browse such file (for linking or appending), the
   FileBrowser shows an error message instead of the content of the
   library file.
* When a newer fileversion blendfile is open, a permanent warning is
  shown in the footer of the UI.
** By clicking on it, user can get more info.
** When trying to save such newer file, a popup asks for confirmation,
   with default option to 'Save As' instead.
** If 'save confirmation' option is enabled, when quiting or opening
   another file, and current one is newer, the 'Save' option is replaced
   by a 'Save As' one.

Technically, the hard check on minfileversion is implemented as a new
util in readfile code, `blo_check_minversion`, which is called as part
of the process to get a FileData (aka BlendHandle) from a file path.

The 'soft warning' about data loss due to forward compatibility
limitations is stored as a new flag in Main,
`has_forward_compatibility_issues`.

NOTE: Final commit to main (and backports to both LTSs) will happen as
several commits.

Implements #109151.