Eevee-next: Use Uniform Buffer for Reflection Probes
Our intent was that uniform buffers would have been used already. The
buffer was already marked as uniform buffer, but still was attached
as an SSBO leading to incorrect bindings.

Until now we didn't see any artifacts, but it would lead to more
swapping in/out of buffers that might not be needed.