Replaced socket type structs with a data pointer inside generic sockets.
The reason is that swapping out the socket type of an item would require
replacing the entire struct, which invalidates pointers and is difficult
to manage in RNA.

The socket_data uses the existing `bNodeSocketValueXXX` structs from
nodes. Eventually nodes might also use this new implementation, and
code could be moved to a dedicated file, so adding socket types becomes
easier (less spread out code, most stuff in one place, better checks
when things are not implemented for a type).