Cleanup: CMake: Modernize bf_rna dependencies
Pretty straightforward
- Remove any makesrna paths from INC
- Add a dependency though LIB

Not so pretty straightforward:

- bf_rna will build after most libs it depends on but some of those
libs need the RNA_prototypes.h header which as far as CMake knows
only bf_rna relies on, so it is free to generate it "just" before
bf_rna builds, which is problematic for everybody else since the
header may not exist on disk yet.

An additional target is introduced for this bf::rna::prototypes that
gives access to this header, and ensures makesdna runs before anything
that relies on this header tries to use it.