Makesdna: more verbose `Align struct error`
Change the error message from:

```
Align struct error: Bone color
```

to:

```
Align struct error: Bone::color (starts at 180 on the native platform;
 180 % 8 = 4 bytes)
```

The colon notation (`Bone::color`) makes it easier to recognise that this
is about a specific struct field. The rest of the explanation makes it
clear that this is about the start of the struct. It includes the math
the check is actually doing, providing concrete information on how to
change the code to fix the issue.