Fix #110255: Cover up CPU/GPU differences in light tree.
This pull request covers up a subtle difference between the CPU and GPU
when rendering with a light tree. Specifically a case where the user
has a sun light, with a small angle.

The difference was caused by the dot() function being different between
CPU and GPU backends, with the GPU showing more floating-point
precision losses.