Fix: shader_builder build
somewhat of an oddball, shader_builder is off by default, but clement
mentioned i broke it, so I went out to fix it (It's on by default now
for me locally while testing any build system changes)

Shader_builder should be a fairly light executable that checks if
shaders compile but, since it depends on bf_gpu, which drags in
blenkernel, which drags in all of blender, it's a bit chunkier than one
would hope it would be.

Either way, one of the cleanups in the last week gave it some build
issues which are easily resolved by adding the missing dependency on
bf_bmesh, and once that's fixed, it gave some linker errors, caused by
link order issues due to missing dependencies inside cycles.

- cycles_util needs to link glog, since log.cpp uses it
- cycles_scene need to link cycles_graph since shader.cpp uses the node
api

These changes make it work for now, but I'm fairly confident this wasn't
the last of the link order issues.