Transform: update transformation while navigating
Previously, while navigating in the 3D View, the object remained in the
transformation prior to navigation, and was only updated when the
navigation ended.

This commit updates the transform during navigation.

This is only noticed for View Rotate/Orbit and Pan with Trackpad or
NDof (not with the mouse since it follows the initial position of the
object).