Modifiers: Add `panelIDName` field
Previously the panel id name of a modifier, e.g. "MOD_PT_Smooth" was
created by copying from the `ModifierTypeInfo`s `name`.
This meant that modifiers with the same default name would use
the same identifier for the panels.

This PR refactors this by introducing `panelIDName` allowing the
modifier to specify the name of the panel independantly of the default
name.