I18n: make favorites menu entry names translations depend on language
Buttons can be added to the quick favorites menu. The name of this
button is stored in a menu entry structure. Depending on the button
type (property, menu, operator), the name will be obtained in
different ways.

The correct name to store to is the untranslated one, because we want
to translate only when the menu is called. But in the case of
operators, the name copied to the struct was translated. This caused
the quick favorites menu to ignore language change and always display
the name as it was translated at the time of addition.

This commit adds a new function to get the raw, untranslated operator
name instead.

Fixes #110368: Favorites menu does not take into account language change

TODO
- menus
- operators with attributes (e.g. add specific mod),
- ...?